+++ szconfigmanager/doc.go
@@ -1,4 +1,4 @@
 /*
-Package szconfigmanager is used to modify Senzing configuration JSON documents in the Senzing datastore.
+Package szconfigmanager is used to modify Senzing configuration JSON documents in the Senzing repository.
 */
 package szconfigmanager